requirements.txt
Django==4.2
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
sentence-transformers==3.0.1
fastapi==0.114.0
uvicorn==0.30.6
pandas
numpy